% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BACdecision.R
\name{plot.BACdecision}
\alias{plot.BACdecision}
\title{Generateing Plot(s) Used for Decision Rule Evaluation}
\usage{
\method{plot}{BACdecision}(x, es.null = 0, es.null.side, interim = F, ...)
}
\arguments{
\item{x}{An object of "BACdecision" class.}

\item{es.null}{Effect size under the null hypothesis. Default is 0.}

\item{es.null.side}{"=" is the only option now.}

\item{interim}{Logical indicator of interim analysis (versus final
analysis). Figures will differ in legends and titles.}

\item{...}{Argument to be passed to or from other methods}
}
\value{
An object of "ggplot" class. Certain further edits are still allowed,
  such as changing title and adding reference lines.
}
\description{
\code{plot} method for class "BACdecision"
}
\details{
If \code{interim=F}, only one power/type I error figure will be
  generated. Otherwise, two figures will be generated correponding to "No Go"
  and "Go" decisions respectively.
}
\author{
Hongtao Zhang
}

