% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BAC.R
\name{summary.BAC}
\alias{summary.BAC}
\title{Summarizing BAC}
\usage{
\method{summary}{BAC}(object, ...)
}
\arguments{
\item{object}{An object of class "BAC"}

\item{...}{Argument to be passed to or from other methods}
}
\description{
\code{summary} method for class "BAC"
}
\author{
Hongtao Zhang
}

